ref: latest block controller query
